refactor(examples): split rust vm smoke harness
Move the configurable filesystem, block, and Windows virtio-console smoke paths out of rust_vm into a dedicated rust_vm_smoke example crate.

Keep rust_vm focused on the simple Rust API hello-world path and wire the new example into the workspace and Makefile.
